feat(game-options): lovers reveal each other role game option
