+++ config/eslint/rules/overrides/factory-files.js
@@ -1,8 +1,16 @@
-const { OFF } = require("../../constants");
+const { OFF, ERROR } = require("../../constants");
 
 const factoryFilesOverride = Object.freeze({
   files: ["*.factory.ts"],
-  rules: { "@typescript-eslint/no-magic-numbers": OFF },
+  rules: {
+    "@typescript-eslint/no-magic-numbers": OFF,
+    "import/max-dependencies": [
+      ERROR, {
+        max: 30,
+        ignoreTypeImports: true,
+      },
+    ],
+  },
 });
 
 module.exports = { factoryFilesOverride };
