+++ src/modules/game/dto/create-game/create-game-options/create-roles-game-options/create-cupid-game-options/create-cupid-lovers-game-options.dto.ts
@@ -1,0 +1,17 @@
+import type { ApiPropertyOptions } from "@nestjs/swagger";
+import { ApiProperty } from "@nestjs/swagger";
+import { IsBoolean, IsOptional } from "class-validator";
+
+import { CUPID_LOVERS_GAME_OPTIONS_API_PROPERTIES, CUPID_LOVERS_GAME_OPTIONS_FIELDS_SPECS } from "@/modules/game/schemas/game-options/roles-game-options/cupid-game-options/cupid-lovers-game-options/cupid-game-options.schema.constant";
+
+class CreateCupidLoversGameOptionsDto {
+  @ApiProperty({
+    ...CUPID_LOVERS_GAME_OPTIONS_API_PROPERTIES.doRevealRoleToEachOther,
+    required: false,
+  } as ApiPropertyOptions)
+  @IsOptional()
+  @IsBoolean()
+  public doRevealRoleToEachOther: boolean = CUPID_LOVERS_GAME_OPTIONS_FIELDS_SPECS.doRevealRoleToEachOther.default;
+}
+
+export { CreateCupidLoversGameOptionsDto };
