+++ tests/unit/specs/modules/game/dto/base/decorators/additional-cards/additional-cards-for-thief-size.decorator.spec.ts
@@ -8,7 +8,7 @@ import { createFakeCreateGameAdditionalCardDto } from "@tests/factories/game/dto
 import { createFakeCreateThiefGameOptionsDto } from "@tests/factories/game/dto/create-game/create-game-options/create-roles-game-options/create-roles-game-options.dto.factory";
 import { createFakeCreateGameDto } from "@tests/factories/game/dto/create-game/create-game.dto.factory";
 import { createFakeGameOptions } from "@tests/factories/game/schemas/game-options/game-options.schema.factory";
-import { createFakeRolesGameOptions } from "@tests/factories/game/schemas/game-options/game-roles-options.schema.factory";
+import { createFakeRolesGameOptions } from "@tests/factories/game/schemas/game-options/game-roles-options/game-roles-options.schema.factory";
 
 describe("Additional Cards For Thief Size Decorator", () => {
   describe("isAdditionalCardsForThiefSizeRespected", () => {
